improve docs
